+++ samples/Java/samples/testbed/EWrapperImpl.java
@@ -49,7 +49,7 @@ public class EWrapperImpl implements EWrapper {
 	
 	//! [ticksize]
 	@Override
-	public void tickSize(int tickerId, int field, int size) {
+	public void tickSize(int tickerId, int field, long size) {
 		System.out.println("Tick Size. Ticker Id:" + tickerId + ", Field: " + field + ", Size: " + size);
 	}
 	//! [ticksize]
@@ -189,7 +189,7 @@ public class EWrapperImpl implements EWrapper {
 	//! [updatemktdepth]
 	@Override
 	public void updateMktDepth(int tickerId, int position, int operation,
-			int side, double price, int size) {
+			int side, double price, long size) {
 		System.out.println("UpdateMarketDepth. "+tickerId+" - Position: "+position+", Operation: "+operation+", Side: "+side+", Price: "+price+", Size: "+size+"");
 	}
 	//! [updatemktdepth]
@@ -197,7 +197,7 @@ public class EWrapperImpl implements EWrapper {
 	//! [updatemktdepthl2]
 	@Override
 	public void updateMktDepthL2(int tickerId, int position,
-			String marketMaker, int operation, int side, double price, int size, boolean isSmartDepth) {
+			String marketMaker, int operation, int side, double price, long size, boolean isSmartDepth) {
 		System.out.println("UpdateMarketDepthL2. "+tickerId+" - Position: "+position+", Operation: "+operation+", Side: "+side+", Price: "+price+", Size: "+size+", isSmartDepth: "+isSmartDepth);
 	}
 	//! [updatemktdepthl2]
@@ -649,7 +649,7 @@ public class EWrapperImpl implements EWrapper {
 
     //! [tickbytickalllast]
    @Override
-    public void tickByTickAllLast(int reqId, int tickType, long time, double price, int size, TickAttribLast tickAttribLast,
+    public void tickByTickAllLast(int reqId, int tickType, long time, double price, long size, TickAttribLast tickAttribLast,
             String exchange, String specialConditions) {
         System.out.println(EWrapperMsgGenerator.tickByTickAllLast(reqId, tickType, time, price, size, tickAttribLast, exchange, specialConditions));
     }
@@ -657,7 +657,7 @@ public class EWrapperImpl implements EWrapper {
 
     //! [tickbytickbidask]
     @Override
-    public void tickByTickBidAsk(int reqId, long time, double bidPrice, double askPrice, int bidSize, int askSize,
+    public void tickByTickBidAsk(int reqId, long time, double bidPrice, double askPrice, long bidSize, long askSize,
             TickAttribBidAsk tickAttribBidAsk) {
         System.out.println(EWrapperMsgGenerator.tickByTickBidAsk(reqId, time, bidPrice, askPrice, bidSize, askSize, tickAttribBidAsk));
     }
