+++ source/csharpclient/TwsRtdServer/TwsRtdServerWrapper.cs
@@ -172,7 +172,7 @@ namespace TwsRtdServer
             SetTopicValue(tickerId, field, price);
         }
 
-        public void tickSize(int tickerId, int field, int size) 
+        public void tickSize(int tickerId, int field, long size) 
         { 
             SetTopicValue(tickerId, field, size);
         }
@@ -228,8 +228,8 @@ namespace TwsRtdServer
         public void historicalDataUpdate(int reqId, Bar bar) { }
         public void historicalDataEnd(int reqId, string start, string end) { }
         public void marketDataType(int reqId, int marketDataType) { }
-        public void updateMktDepth(int tickerId, int position, int operation, int side, double price, int size) { }
-        public void updateMktDepthL2(int tickerId, int position, string marketMaker, int operation, int side, double price, int size, bool isSmartDepth) { }
+        public void updateMktDepth(int tickerId, int position, int operation, int side, double price, long size) { }
+        public void updateMktDepthL2(int tickerId, int position, string marketMaker, int operation, int side, double price, long size, bool isSmartDepth) { }
         public void updateNewsBulletin(int msgId, int msgType, String message, String origExchange) { }
         public void position(string account, Contract contract, double pos, double avgCost) { }
         public void positionEnd() { }
@@ -283,8 +283,8 @@ namespace TwsRtdServer
         public void historicalTicks(int reqId, HistoricalTick[] ticks, bool done) { }
         public void historicalTicksBidAsk(int reqId, HistoricalTickBidAsk[] ticks, bool done) { }
         public void historicalTicksLast(int reqId, HistoricalTickLast[] ticks, bool done) { }
-        public void tickByTickAllLast(int reqId, int tickType, long time, double price, int size, TickAttribLast tickAttribLast, string exchange, string specialConditions) { }
-        public void tickByTickBidAsk(int reqId, long time, double bidPrice, double askPrice, int bidSize, int askSize, TickAttribBidAsk tickAttribBidAsk) { }
+        public void tickByTickAllLast(int reqId, int tickType, long time, double price, long size, TickAttribLast tickAttribLast, string exchange, string specialConditions) { }
+        public void tickByTickBidAsk(int reqId, long time, double bidPrice, double askPrice, long bidSize, long askSize, TickAttribBidAsk tickAttribBidAsk) { }
         public void tickByTickMidPoint(int reqId, long time, double midPoint) { }
         public void orderBound(long orderId, int apiClientId, int apiOrderId) { }
         public void completedOrder(Contract contract, Order order, OrderState orderState) { }
